Refactor cypress tests for editing ASN
